Move `POSEntryPointController` to a separate file.
