+++ WooCommerce/Classes/POS/Controllers/POSEntryPointController.swift
@@ -1,0 +1,29 @@
+import SwiftUI
+import protocol Experiments.FeatureFlagService
+
+@available(iOS 17.0, *)
+@Observable final class POSEntryPointController {
+    private(set) var eligibilityState: POSEligibilityState?
+    private let posEligibilityChecker: POSEntryPointEligibilityCheckerProtocol
+    private let featureFlagService: FeatureFlagService
+
+    init(eligibilityChecker: POSEntryPointEligibilityCheckerProtocol,
+         featureFlagService: FeatureFlagService = ServiceLocator.featureFlagService) {
+        self.posEligibilityChecker = eligibilityChecker
+        self.featureFlagService = featureFlagService
+
+        guard featureFlagService.isFeatureFlagEnabled(.pointOfSaleAsATabi2) else {
+            self.eligibilityState = .eligible
+            return
+        }
+        Task { @MainActor in
+            eligibilityState = await posEligibilityChecker.checkEligibility()
+        }
+    }
+
+    @MainActor
+    func refreshEligibility() async throws {
+        // TODO: WOOMOB-720 - reason based eligibility refresh
+        eligibilityState = await posEligibilityChecker.checkEligibility()
+    }
+}
